pyproject.toml
[project]
name = "transf-arch-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "loguru>=0.7.3",
    "polars>=1.34.0",
    "pydantic>=2.12.3",
    "pylint>=4.0.1",
    "requests>=2.32.5",
    "ruff>=0.14.1",
    "tokenizer>=3.5.3",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
